take the `Source` code in its own file
Take the `Source` code out of mod.rs and into source.rs. Following
commits will add code over there.
